configure: add extra code checks matching GCC 14
In gcc 14 some warnings are now turned into errors by default [1].
So of these were already errors in VLC. But there's are some that we can use.

This will ensure the VLC code is ready for the GCC 14 move.
However there's a chance some contribs don't build anymore, especially the old
unmaintained ones.

[1] https://gcc.gnu.org/gcc-14/porting_to.html#warnings-as-errors