qt: detect Windows 11 22H2 properly in win32windoweffects module
Windows 11 identifies as Windows 10 (NT 10),
so this comparison always yields false.

As a side note, native acrylic background
for Win32 windows only gained support with
the Windows 11 22H2 update, that's why
there is the check here.

Note that according to the general conventions
a transparent window must be frameless on
Windows. This is currently not the case,
because "frameless window don't support
areo snapping". So, it can be considered
as a known bug if native acrylic does not
work as intended. This grants a requirement
of change to use frameless window, and get
rid of the current workaround that simulates
frameless window in mainctx_win32.cpp. If
aero snapping does not work, it is Windows's
problem since we are using frameless windows
in every other platform.