VLCMediaSource: check file:// for file browsing
An INPUT_TYPE_DIRECTORY might not match with an MRL supported by the
CoreFoundation API. The code handling the file browsing is made for
local file browsing at locations that are not exposed by services
discovery.

Ensure file:// is the scheme used for the MRL to trigger this browsing
code. This is not 100% compliant since it doesn't account for a demux
but there's probably no reason to have a demux there, and no occurrence
to that day.

This avoid bypassing the vlc_media_source/media_tree preparsing when the
source has been reported as INPUT_TYPE_DIRECTORY.