Revert "configure: generate an error in extra checks when return type is bogus"
This reverts commit 4cd7d490402b1e733f41a5b1b593077a3bc39ab1.

This warning can occur in callbacks that never return. Some functions cannot
have a return statement because that would cause dead code warnings, while
they also cannot return void because the callback prototype is externally
mandated. Ideally the compiler would notice that the function does not
return, and sometimes it does, but evidently we cannot rely on that, so
we have to revert this warning back to a warning, not an error.

Fixes #28650.