qt: do not use double precision in settings
The docs state that:

> Note that INI files lose the distinction
between numeric data and the strings used
to encode them, so values written as numbers
shall be read back as QString. The numeric
value can be recovered using QString::toInt(),
QString::toDouble() and related functions.

Currently, double-precision numbers can not
be reconstructed properly due to QString's
way of handling dp numbers.

Since they are read back as QString, we
should save these numbers according to the
precision that QString is comfortable with,
so that they are parsed properly when the
application starts.