modules: Makefile.am: generate list of plugins
The build system is generating plugins and installing them at the PREFIX
location, but it is also often used by extra build scripts using it.
Those build scripts were trying to list the plugins by listing the files
in the PREFIX folder, but it was often inadequately using stall
plugins.

By generating an up-to-date list of plugins by the build system, we can
check when the plugin list have changed and which plugins should be
copied to the final application.