extras: apple: add extra version for simulators
The simulators are constrained to specific iOS and tvOS minimal version
and not respecting the constraint would lead to incorrect build and
linking failure.

Not sure how this should really be handled yet though, maybe the
build.sh script should get some parameter for that, or define the
targets depending on the SDK version.