qml: use image for the shadow in Player instead of shader effect
`DropShadow` available in `Qt5Compat` does not support adjusting
samples (Qt 6.2), because we do not want to compile shaders at
run-time (explicit opt-out on Qt 6.7).

The default samples setting yields an inferior shadow quality
compared to the pre-generated shadow image. Not only that,
depending on the platform, it may appear completely bugged.

Here, this shadow is a stand-alone shadow so we do not need
caching. Without caching, using pre-generated image can be
justified also considering the (a)synchronoucity.