sout: hls: fix invalid playlist ref on error path
The playlist reference should only be held in the variant map when it's
asserted that the playlist creation will succeed. Otherwise the error
handling will delete the playlist while keeping an invalid reference in
the map.
The reference is already properly stored at the end of the `Add`
function, so this patch only reworks an unnecessary store.