CI: build LLVM Windows nightly builds with PDB files
The PDB files will be inside the -debug.7z package.

The PDB source mapping will be the '<root>/vlc' folder to '/'.

In Visual Studio you can map to your local source folder with:
> Solution > Properties > Debug Source Files > "s:/sources/vlc"

In Visual Studio Code you can map to your local source folder with the
"sourceFileMap" variable in the launch configuration
https://code.visualstudio.com/docs/cpp/launch-json-reference#_sourcefilemap:

> "sourceFileMap": { "/": "s:/sources/vlc" }