filter_chain: fix activation function type
Before the previous commit, the activation function was implicitely of
type int (*)(vlc_object_t *) because module_need was used. But the 
following commits changed the activation function for the modules:

 - 95c4582681a994a5c2a9301b20015151eec7a101
 - c9f42135fb5c9eeb740a522116036707ce921eb7
 - f9ab141d20fe03fd06fbb8ab2f08849e9c35afe3
 - b027f1a6c792da764bec4290b2510bbc14dc18a3
 - 94e23d51bb91cff1c14ef1079193920f04f48fd1
 - 0d5d447d704644f425727a86404ba1646b06062a
 - 1c87e2904259201f0c0b08c44a30e38003c36848

It led to the activation function being called with non-matching
parameters, which is UB and crashes on some platforms where types are
checked, like when executing from WASM VMs.

This commit explicitely changes the type to a matching one for every
filter type.