qt: add setting to hide the window on close
When `qt-close-to-system-tray` is set,
the application would not exit on close
event, but rather the window becomes
hidden.

This setting is not enabled by default,
and is only applicable when `qt-system-tray`
is also set.

I believe that with the media library,
this setting becomes relevant to keep
one instance open rather than closing
and re-opening. Currently, even if the
user does not want to close the application,
the inclination is to close the application
since the window occupies space in the
task bar.