rust: add vlcrs-macros crate for module!{} macro
This macro tries to do the same things as the one in C but in a Rust-ty
way.

Signed-off-by: Alexandre Janniaux <ajanni@videolabs.io>

Picked from main rust MR as-is, so it won't be compiling correctly.
Follow-up changes will add the crates to the workspace members after
adapting to the new tree.

The crate will provide the module!{} macro to define the plugin manifest
in a declarative and type-safe way, bringing the language part to create
new modules in Rust. Only the buildsystem part will be missing after
this crate is enabled.