vlcrs-macros: fix for-loops used as if
cargo clippy was complaining that the loop would trigger only once.
Indeed, the for-loop is used to trigger iter() and return immediately
with an error at the first element.

Using if-let deconstructuring construct on the first() element achieve
the same feature while providing more semantics on what it is checking.