vlcrs-macros: use capability trait to load modules
The previous implementation from Loïc Branstett was using public
activation and deactivation functions that would be fetched from the
capability name. But it meant that capabilities could not have spaces in
them, removing the possibility to use "video filter" or requiring to
patch the name beforehand, and it was fit to a model with a single crate
so as to know where to find the activation functions.

Since the introduction of the binding generator in commit
bd1aaaf4ed5dc89761b0b65b24a31316341c830c, our crates are separated and
sorted to match each include file from vlccore.

In addition, it meant that it would not have been possible to write
out-of-tree capabilities while enjoying the type-safe features brought
by the bindings, as opposed to the current C API. Finally, it was also
not compatible with custom loading function type.

This commit brings a new indirection by allowing capability traits to
define `Loader`, `Activate` and `Deactivate` types in their definitions,
in an effort to keep a type-safe interface when __writing__ modules, as
well as being able to easily describe new module interfaces without
compromising on the current features of the module infrastructure.

       Module Implementor                      Core support

    - Implement capability              - Provides capability traits 
      traits                            - Provides module loader
                                          functions

The types `Activate` and `Deactivate` would allow to define custom
activation function type, such as those we find when using
vlc_module_load or manually probing using vlc_module_match, as opposed
to the common vlc_object_t-based module_need() activation function.

The type `Loader` allows to define a default associated type that would
provide the activation functions to the module manifest writer, so that
the correct interface is used. It effectively allows capability traits
to define custom activation function types, without having the module
implementor explicitely provide those functions. The module implementor
is still able to provide their own activation functions by providing
their own loader.