CI: rework Darwin jobs
This updates the ARM64 job for macOS to use the latest SDK (14). The
existing x86_64 job is kept unchanged.

The tvOS job is updated to use tvOS 17.

The iOS ARM64 and x86_64 simulator jobs are updated to iOS 17.

The current iOS 15 armv7 job is kept.

All that is possible by adding a new runner with Xcode 15.4 on macOS
14.5 with a M1 architecture.