vlcrs-macros: add test for module!{}
The tests are checking:
 - [x] whether the module manifest has wanted opcode generated
 - [ ] whether the module manifest is 100% correct
 - [ ] whether the arguments for the module manifest opcode are correct
 - [x] whether we can create modules with usual activation functions
 - [x] whether we can create modules with custom activation functions
 - [x] whether we can omit the deactivation function
 - [x] whether the activation function is correctly called
 - [x] whether the deactivation function is correctly called
 - [x] whether the infrastructure works for submodules

The test module.rs provides a simple test with normal interface to check
the module manifest with open and close functions.

The test module_multiple.rs is important to ensure the macro doesn't
generate code with duplicated identifiers, that would trigger errors
only when multiple modules are declared from a single manifest.

The test module_default.rs checks the minimal definition required for
core module loading implementation to define a new capability, using
the default functions as much as possible.

The test module_specific checks that we can use other signatures and
uses this feature to check that the activation and deactivation
functions are correctly called when using the module API.