vlcrs-macros: add capability trait in module!{}
The capability trait will embed the type system which brings the
activation and deactivation function for the module, as well as some
future changes regarding capability checking at compile time.

Forcing to specify the capability trait enables a struct to be
implemented by multiple capabilities, and enable the module manifest
macro to know which capability trait is actually associated with which
capability specification, and will enable static checks to be performed
at compile time.