vlcrs-plugin: add ModuleProtocol trait
The trait defines a protocol to be implemented by any "Loader" for the
"CapabilityTrait". Both terms will be defined in follow-up commits.

In particular, this trait enforce the module loaders to provide the
activation function and deactivation function, as well as providing
default values for the deactivation functions.

More in-depth experiments were done to extend this approach to the
capability trait, so that traits defining a capability and traits
defining the ModuleLoader interface are separated, but automatically
implementing a capability trait for module structures that would use
the capability interface requires the following kind of declaration:

    impl<T : SpecificCapabilityTrait> CapabilityProtocol for T
        where T : SpecificCapabilityTrait
    {
        type Activate = vlc_activate;
        type Deactivate = vlc_deactivate;
        type Loader = ModuleLoader<Self>;
    }

In this case, the implementation of the trait is done in crates exposing
the capability, for instance vlcrs-interface or an external crate
exposing an out-of-tree module, and the CapabilityProtocol is exposed by
the vlcrs-plugin crate, so it is considered as a generic implementation
of a foreign type, which is not allowed and would end up with the
following error:

    error[E0210]: type parameter `T` must be used as the type parameter for some local type (e.g., `MyStruct<T>`)
      --> vlcrs-macros/tests/module.rs:52:6
       |
    52 | impl<T : SpecificCapabilityModule> CapabilityProtocol for T
       |      ^ type parameter `T` must be used as the type parameter for some local type
       |
       = note: implementing a foreign trait is only possible if at least one of the types for which it is implemented is local
       = note: only traits defined in the current crate can be implemented for a type parameter

    Some errors have detailed explanations: E0210, E0576.
    For more information about an error, try `rustc --explain E0210`.