vlcrs-macros: adapt documentation to changes on traits
The module!{} macro documentation should now expose an existing
dedicated capability trait and specify the loader. We don't want that to
be visible from the module!{} macro, so everything is hidden, but we
might want to replace the specific capability by an exisiting capability
trait when some will exist in the future, so that the documentation is
simplified and the example matches real life usage.

Otherwise, the doctest could be turned into `ignore` instead of
`no_build` but it means that the documentation could become out-of-date
with what the users are supposed to do.