filter: add a function to load filters
The activate function should match exactly vlc_filter_open, rather than use
a generic vlc_object_t loader. Then we don't have any dirty cast.

The code is similar to module_need, with better typing.

It's not possible to inline this because we need to call vlc_objres_clear()
for each module loaded unsuccessfully. It's not a public API.