vlc_filter: use a single type to open all types of filters
Because of the filter chain, they must have a common way to open them.

Some types of filters may not be used by the filter chain (for example the text
renderer). But still need to be activated the same way. If this ever changes
They will need their own open callback type back.