contrib: rust: split cargo-c and rust+cargo
They are not the same thing. cargo-c is a rust crate.
It requires rustc and cargo to be installed. It doesn't need
any package download via the contrib downloads.

rustc and cargo also don't need any download via the contrib downloads if
rustup is already installed. It's the one that will do all the downloads.