qt: add preference `qt-backdrop-blur`
This setting makes it possible to disable the
backdrop blur filter for the main window.

Apart from accessibility reasons, having this
setting enabled makes text being rendered in
worse quality since Qt's sub-pixel rendering
is impaired by translucency, which is a
requirement for exposing backdrop effects in
the UI layer.

In order to align with the proposed design,
this setting is enabled by default.