Revert "contrib: rav1e: do the vendoring untar/build during the build phase"
This reverts commit 325123084b2529634bb5e203416712e21f91bf25.

The optional dependency doesn't work.
But in the end we can just create the tarball right after the download failed.
So the rav1e-$(RAV1E_VERSION)-vendor.tar.bz2 is guaranteed to create a usable tarball (or the build will fail).

Ultimately it should be the job of the download_vendor macro to create it if the download fails.