demux: mkv: add a template to get mandatory elements
We get a reference to the object, or it throws an error that a mandatory element
is missing.

This should only be used for elements without a default value. Otherwise the
missing element would be added automatically to get the default value. But
since it's a const EbmlElement, we don't want to modify it.