contrib: rav1e: require rustc+cargo before we get the vendor tarball
Since we want to build the vendor tarball if it's not there, we require
rustc+cargo to be available when calling rav1e-vendor-build.
However we cannot do the dependency in the rav1e-vendor-build target
since it would be handled in another process that handles the contrib dependencies.
This can lead to a race condition on building rust+cargo or rustup from
different MAKE processes.

That means we will download the native cargo+rustc during the fetch stage.
Which is actually a good thing, except it won't be in our tarballs.

Co-authored-by: Martin Storsjö <martin@martin.st>