clock: implement vlc_clock_Start
This commit creates a new way in the clock system to intercept the
behaviour of the "startup" depending on whether the clock is for the
input or the output and whether it's a reference clock or a derived one.
Allowing to specify this behaviour contributes to unify the usage of the
clock, so that it's easier to use in the future (think start() with the
current time and startup PCR as planned vs SetFirstPCR() with a
corrected startup time and the last PCR received). It also contributes
to improve the cases where the input is a reference clock and paced by
the source.

The es_out is handling the current startup sequence for the clock by
resetting the clock, shifting the input clock and setting the startup
time again. Those functions are mixing three different components and
behaviour and makes it hard to test the startup scenarios.

vlc_clock_Start() on the input vlc_clock_t will replace those calls by
providing the correct offset on the output points instead of
transforming the input points.

In addition, vlc_clock_Start() will be provided to the output clocks so
that they can provide the wait_sync reference point, instead of relying
on the first conversion done, and specify the exact startup time for the
output clocks by computing the additional delay at the moment the
playback really starts.

Like first_pcr, start_time must be shifted when pause is set. It would
also be possible to "reset" the start_time when un-pausing to the
un-pause date.

Note that this commit doesn't change the behaviour yet, and the
consequences it brings will be seen when SetFirstPCR is effectively
removed. Current case coverage is also quite low in tests, so failures
would happen without SetFirstPCR in a random way.

Co-authored-by: Thomas Guillem <thomas@gllm.fr>