clock: setup new clock_context selection
Ensure that every clocks (even input) has a valid clock_context, and
specifically select them during vlc_clock_Start(), and un-select them
during vlc_clock_Reset().

The previous mode is still present in the code to avoid breaking the
current code until all the necessary changes are introduced, and is
selected by whether the main_clock->first_pcr value has been set or not.

Now that vlc_clock_main_Reset() is not called anymore at this location,
the points registered from the input when it is driving the clock bus
will not be removed.

    /!\ As a consequence, --clock-master=input now starts from the input
    /!\ clock points instead of starting from a monotonic clock point.

This indirectly fixes an issue where the output_dejitter is accounted in
computing the wait_sync_ref date, but also making all the outputs late
compared to the input clock, leading to "Deferring start" messages from
the audio output, and then "Starting late" message.

Co-authored-by: Thomas Guillem <thomas@gllm.fr>