codec: videotoolbox: fix incorrect format display size
The input sample's CMVideoFormatDescription needs to be created with
visible pixels, not the buffer dimensions. Not doing that causes the
output image buffer to report incorrect visible size when queried with
CVImageBufferGetDisplaySize.

This can cause issues on display/encoding paths that solely rely on
CVBuffer metadata, like AVSampleBufferDisplayLayer or VideoToolbox CVPX
encoding.

Fix #28703