input_clock: don't update invalid points on rate change
Only call UpdateListener() if the clock is valid.

This fixes the following assert:

vlc: ../../src/clock/clock.c:585: vlc_clock_monotonic_to_system: Assertion `ctx->wait_sync_ref.stream != VLC_TICK_INVALID' failed.

This was occurring after a pause, a flush and a resume (that triggers a
rate change).

The assertion is new, but VLC_TICK_INVALID was used as a reference point
before in that case.