mux: asf: header: set correct number of sub objects
The Microsoft ASF spec states

```
Number of Header Objects
Specifies the number of objects contained within the Header Object, not
including this one. In addition, the Header Extension Object (sections 3.4
and 4) is counted as exactly one object regardless of the number of
subobjects contained within it.
```

This means `i_cm_size` should not contribute to the number of sub objects.
As a result VLC generated ASF files could not be demuxed by gstreamer's
`asfdemux`.

Fix the number and while at it give some semantic meaning to the magic
numbers in the code regarding the header sub object.

Though technically we write an unsigned value use int to stay consistent
with the other size values

Closes #28714

Signed-off-by: Sebastian Gross <sgross@emlix.com>