contrib: rust: skip hash check for generated vendor tarball
We will never have the same tarball on all platforms using
different versions of tar, different order of files, different owner, etc.

Since we generated the tarball we don't have to verify
the checksum, we can assume it's OK, assuming cargo vendor is safe.

If we ever delete the tarball, the marker is deleted so we check
the download properly from the download if possible.