contrib: Fix the check for _UCRT
_UCRT is not a compiler internal define. Some build scripts may be
invoking the compiler with an explicit -D_UCRT, but if not, the
toolchain default is set by the toolchain headers. Therefore, include
crtdefs.h to pull in default headers for checking whether it is set
or not.

Also check for older forms of mingw-w64, which didn't use the
define _UCRT to signal the type of CRT.