vout: apple: Add support for video orientations in SampleBufferDisplay
Transform image buffer's pixels to display frames with proper video orientation.
API used to transform the picture depends on the host operating system version.
Given OS is at least iOS/iPadOS/tvOS 16.0 or macOS 13.0, an implementation backed
by VideoToolbox APIs will be used.
If not available, it will fallback to a CoreImage backed implementation.
So far the CoreImage implementation doesn't seems to perform as fast as the one
backed by the VideoToolbox API.