contrib: rust: allow both MSVCRT and UCRT with windows-gnu*
When using a static Rust libstd the linker can pick the runtime it wants.
It will correspond to the same runtime we use for the rest of the code.

See https://code.videolan.org/videolan/vlc/-/merge_requests/5675#note_447873

gnullvm targets most importantly use LLVM libraries and ABI. So libgcc is replaced
with libunwind, the linker is LLD, it uses native TLS instead of emutls, function
sections are enabled, and there's probably something else that I forgot. UCRT
is another difference, but a smaller one since the changes from my previous
sentence cause builds with GCC/Binutils to break.

windows-gnu targets are built with MSVCRT, but that shouldn't be a big deal as
long as you stick to the default static std library (in other words, as long as
you don't use -C prefer-dynamic). I can't guarantee it will always work, but
t's much better compared to using windows-gnullvm targets, which have a 100%
chance of failing when using them with GCC/Binutils. Unless you want to migrate
to LLVM as well (windows-gnullvm targets by default will pick Clang from PATH
as the linker), but that would be a much bigger change than this PR implies.