qt: fix video surface initialization and de-initialization ordering issues
I wanted to initialize video surface provider and
window handler within `commonSetupVoutWindow()`, since
it is where the video window is set and video surface
provider and window handler only makes sense when
there is a valid vout window. However, for some reason,
it did not work so it is a TODO for now.

Currently, since there are a lot of signal connections,
changes are reported even when there is no valid vout
window. For example, `CompositorVideo::onSurfacePositionChanged()`
may be called when there is no vout window or when it
is already invalidated.