qt: fix lua interface extension not loading
extension_dialog_t is not a registered type in qt so the signal was failing.

It's usually dangerous to pass VLC C object in Qt signals regarding the lifetime
of these objects, so rather than registering the type, the object is passed
though lambda capture assumming that the current comment and code is correct.