input_clock: finer backward discontinuity detection
This commit fixes a regression from
03d91093af3fb6f1bd913715bc4656b87ead8f09 that prevented a RESET_PCR when
the input was more than 300ms late.

Indeed, a PCR going backward must be a discontinuity, there is no need
to compare it with system dates. Comparing with system dates could lead
to a backward system discontinuity, that is generally handled by
increasing the pts-delay and calling RESET_PCR (when
input_clock_Update() return late > 0).