input: call SetStopStart() for all input threads
Move the SetStopStart function to ensure it is called for all input threads, especially
during preparsing. It ensures that the correct timings are used before the duration of
the input_item_t is set. Previously, 'start-time' and 'stop-time' values were not parsed
during preparation, leading to priv->i_start and priv->i_stop defaulting to 0.
Refs #4143