avsb: move _timeQueue waiting
It will be faster to wait at the end of the function after we signalled
all components that they need to stop.

It is possible since whenTimeObserved() does not require any others
components that are stopping.

The downside is that a timing has more chance to be reported while
stopping, but it is legal (it should not be reported after stop).