qt: use deep comparison for control bar profile selection
... and promote DEFAULT_STYLE from being a pseudo style
to an actual style.

Addressable types that can not behave like a value type,
such as having an external "ID" field, bring considerable
problems and hinder maintenance. It becomes harder to
serialize which is a necessity since we want to save the
toolbar layout, and new tricky questions arise such as:
are two models that have exactly the same layout and
elements the "same", even if their IDs differ?

This partially reverts commit 09ca278c449b9fac7f90225971c5deb01a354e76,
and uses deep comparison instead of associating ID to the
model for selection purposes.