qt: do not delete QObject from potentially different thread
In 2b0471c, a change was made to delete video surface provider
and video window handler when video window is destroyed. This
makes sense, since they are not necessary anymore (so no need
to keep them in memory) and also the connections would cease
to exist automatically.

However, turns out that this method may be called from a
different thread than the one these objects belong to.
Considering the following quote from Qt documentation, do not
delete them here but instead simply disconnect the signals.

> You must not delete the QObject directly if it exists in a
> different thread than the one currently executing.