qml: consolidate `ListViewExt`'s drag and drop functionality over `TableViewExt`
This makes it possible to not re-invent drop indicator positioning, and also
allows rejecting the drag event when drop is not applicable.

In order to reject drag event when not applicable, there needs to be two
drop areas. This has been the case in `PlaylistListView/Delegate` but not
in `TableViewExt/Delegate`. With this patch, I followed the same approach
used in `PlaylistListView` which has a rather proven track record on drag
and drop functionality.